requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
six==1.16.0
tk==0.1.0
